requirements.txt
absl-py @ file:///tmp/build/80754af9/absl-py_1623867049735/work
astor==0.8.1
certifi==2021.5.30
cloudpickle @ file:///tmp/build/80754af9/cloudpickle_1598884132938/work
cycler==0.10.0
cytoolz==0.11.0
dask @ file:///tmp/build/80754af9/dask-core_1615055117017/work
decorator @ file:///tmp/build/80754af9/decorator_1621259047763/work
gast @ file:///tmp/build/80754af9/gast_1597433534803/work
google-pasta==0.2.0
grpcio==1.14.1
h5py==2.8.0
imageio @ file:///tmp/build/80754af9/imageio_1617700267927/work
importlib-metadata @ file:///tmp/build/80754af9/importlib-metadata_1617877310517/work
joblib @ file:///tmp/build/80754af9/joblib_1613502643832/work
Keras==2.2.4
Keras-Applications @ file:///tmp/build/80754af9/keras-applications_1594366238411/work
Keras-Preprocessing @ file:///tmp/build/80754af9/keras-preprocessing_1612283640596/work
kiwisolver @ file:///tmp/build/80754af9/kiwisolver_1612282412546/work
Markdown @ file:///tmp/build/80754af9/markdown_1614363833670/work
matplotlib @ file:///tmp/build/80754af9/matplotlib-suite_1613407855456/work
mkl-fft==1.3.0
mkl-random==1.1.1
mkl-service==2.3.0
networkx @ file:///tmp/build/80754af9/networkx_1598376031484/work
numpy @ file:///tmp/build/80754af9/numpy_and_numpy_base_1603487797006/work
olefile==0.46
Pillow @ file:///tmp/build/80754af9/pillow_1617386154532/work
protobuf==3.14.0
pydot @ file:///tmp/build/80754af9/pydot_1615393369918/work
pyparsing @ file:///home/linux1/recipes/ci/pyparsing_1610983426697/work
python-dateutil @ file:///home/ktietz/src/ci/python-dateutil_1611928101742/work
PyWavelets @ file:///tmp/build/80754af9/pywavelets_1601658315997/work
PyYAML==5.4.1
scikit-image==0.17.2
scikit-learn @ file:///tmp/build/80754af9/scikit-learn_1621365798935/work
scipy @ file:///tmp/build/80754af9/scipy_1597686625380/work
six @ file:///tmp/build/80754af9/six_1623709665295/work
tensorboard==1.14.0
tensorflow==1.14.0
tensorflow-estimator==1.14.0
termcolor==1.1.0
threadpoolctl @ file:///tmp/tmp9twdgx9k/threadpoolctl-2.1.0-py3-none-any.whl
tifffile==2020.10.1
toolz @ file:///home/linux1/recipes/ci/toolz_1610987900194/work
tornado @ file:///tmp/build/80754af9/tornado_1606942266872/work
typing-extensions @ file:///home/ktietz/src/ci_mi/typing_extensions_1612808209620/work
Werkzeug @ file:///home/ktietz/src/ci/werkzeug_1611932622770/work
wrapt==1.12.1
zipp @ file:///tmp/build/80754af9/zipp_1615904174917/work
